requirements.txt
boto3==1.26.70
pandas<=1.5.1 
awswrangler==2.19.0
pandera==0.13.4
PyYAML==6.0